fix: scroll to the top of the page gets stuck
I recently applied virtual scrolling to the libraries list. I then
noticed some issues with page scrolling when filters were applied.
Sometimes, scrolling to the top of the page didn't work. It got stuck.
So, I decided to replace the native window.scrollTo method with the
index scrolling method from the TanStack Virtual library. The only issue
is that it doesn't use smooth scrolling behaviour due to the different
heights of the items.